Add module and process subsystems.
As the process subsystem needs to make hypercalls the hypercall initializtion
has moved into DriverEntry(). This is ok as it will only ever be called if
the xen driver is not going to conflict with other PV drivers (otherwise
installation would have been vetoed by the co-installer before the service
was added).
Other stuff was moved fixed and changed as I went along; perticularely the
PDO naming code seemed to be wrong.

+++ src/coinst/coinst.c
@@ -1312,7 +1312,7 @@ __DifInstallPostProcess(
     HRESULT                         Error;
     PTCHAR                          DeviceInstance;
     PTCHAR                          ActiveDeviceInstance;
-    WORD                            DeviceId;
+    DWORD                           DeviceId;
     BOOLEAN                         Active;
     BOOLEAN                         Success;
 
@@ -1348,7 +1348,7 @@ __DifInstallPostProcess(
 
     Success = SetFriendlyName(DeviceInfoSet,
                               DeviceInfoData,
-                              DeviceId,
+                              (WORD)DeviceId,
                               Active);
     if (!Success)
         goto fail5;
